+++ lib/types.ts
@@ -37,6 +37,7 @@ export interface CookieOptions {
   domain?: string;
   path?: string;
   secure?: boolean;
+  sameSite?: string
 }
 
 export interface MetricsOptions {
@@ -46,7 +47,7 @@ export interface MetricsOptions {
   maxQueueSize?: number;
 }
 
-interface StoreOptions {
+export interface StoreOptions {
   enabled?: boolean;
 }