+++ test/analytics.test.ts
@@ -11,6 +11,7 @@ var pageDefaults = require('../build/pageDefaults');
 var sinon = require('sinon');
 var tick = require('next-tick');
 var trigger = require('compat-trigger-event');
+
 var Identify = Facade.Identify;
 var cookie = Analytics.cookie;
 var group = analytics.group();
@@ -828,6 +829,22 @@ describe('Analytics', function() {
       });
     });
 
+    it('should should not overwrite context.page values due to an existing bug', function() {
+      analytics.page({ prop: true, context: { page: { search: "lol" } } }, { context: { page: { search: "" } } });
+      var page = analytics._invoke.args[0][1];
+      // FIXME: This assert should fail once the bug is fixed
+      assert.notDeepEqual(page.context(), {
+        page: {
+          ...defaults,
+          search: "lol",
+        },
+      });
+      // FIXME: This assert should fail once the bug is fixed
+      assert.deepEqual(page.context(), {
+        page: defaults,
+      });
+    });
+
     it('should emit page', function(done) {
       analytics.once('page', function(category, name, props, opts) {
         assert(category === 'category');
@@ -1606,6 +1623,29 @@ describe('Analytics', function() {
       analytics.track('event');
       var track = analytics._invoke.args[0][1];
       assert.deepEqual(track.context(), { page: contextPage });
+    });
+
+    it('should support overwriting context.page fields', function() {
+      analytics.track(
+        'event',
+        {},
+        {
+          context: {
+            page: {
+              title: 'lol'
+            }
+          }
+        }
+      );
+
+      var track = analytics._invoke.args[0][1];
+      assert.deepEqual(
+        track.context().page,
+        {
+          ...contextPage,
+          title: 'lol'
+        }
+      );
     });
 
     it('should accept context.traits', function() {